add hardware implementation
