+++ src-tauri/Cargo.toml
@@ -75,6 +75,7 @@ base64 = "0.22.1"
 libloading = "0.8.7"
 thiserror = "2.0.12"
 tauri-plugin-llamacpp = { path = "./plugins/tauri-plugin-llamacpp" }
+tauri-plugin-hardware = { path = "./plugins/tauri-plugin-hardware" }
 
 [dependencies.tauri]
 version = "2.5.0"
