+++ src-tauri/plugins/tauri-plugin-hardware/build.rs
@@ -1,0 +1,5 @@
+use tauri_plugin::Builder;
+
+fn main() {
+    Builder::new(&["get_system_info", "get_system_usage"]).build();
+}
