+++ src-tauri/plugins/tauri-plugin-hardware/permissions/autogenerated/commands/get_system_info.toml
@@ -1,0 +1,13 @@
+# Automatically generated - DO NOT EDIT!
+
+"$schema" = "../../schemas/schema.json"
+
+[[permission]]
+identifier = "allow-get-system-info"
+description = "Enables the get_system_info command without any pre-configured scope."
+commands.allow = ["get_system_info"]
+
+[[permission]]
+identifier = "deny-get-system-info"
+description = "Denies the get_system_info command without any pre-configured scope."
+commands.deny = ["get_system_info"]
