+++ src-tauri/plugins/tauri-plugin-hardware/permissions/autogenerated/commands/get_system_usage.toml
@@ -1,0 +1,13 @@
+# Automatically generated - DO NOT EDIT!
+
+"$schema" = "../../schemas/schema.json"
+
+[[permission]]
+identifier = "allow-get-system-usage"
+description = "Enables the get_system_usage command without any pre-configured scope."
+commands.allow = ["get_system_usage"]
+
+[[permission]]
+identifier = "deny-get-system-usage"
+description = "Denies the get_system_usage command without any pre-configured scope."
+commands.deny = ["get_system_usage"]
