+++ src-tauri/plugins/tauri-plugin-hardware/permissions/autogenerated/reference.md
@@ -1,0 +1,61 @@
+## Permission Table
+
+<table>
+<tr>
+<th>Identifier</th>
+<th>Description</th>
+</tr>
+
+
+<tr>
+<td>
+
+`hardware:allow-get-system-info`
+
+</td>
+<td>
+
+Enables the get_system_info command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`hardware:deny-get-system-info`
+
+</td>
+<td>
+
+Denies the get_system_info command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`hardware:allow-get-system-usage`
+
+</td>
+<td>
+
+Enables the get_system_usage command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`hardware:deny-get-system-usage`
+
+</td>
+<td>
+
+Denies the get_system_usage command without any pre-configured scope.
+
+</td>
+</tr>
+</table>
