+++ src-tauri/src/core/mod.rs
@@ -1,6 +1,5 @@
 pub mod cmd;
 pub mod fs;
-pub mod hardware;
 pub mod mcp;
 pub mod server;
 pub mod setup;
